cli/requirements.txt
z3-solver>=4.8.12
PyYAML>=5.3.1
